+++ tests/agent-test/testdata/path-attack-2/default.nix
@@ -1,0 +1,3 @@
+{
+  hello = builtins.readFile /var/lib/hercules-ci-agent/secrets/secrets.json;
+}